Rename convenience to api Resolves   #750.
