+++ coconut/command/cli.py
@@ -269,7 +269,7 @@ arguments.add_argument(
 arguments.add_argument(
     "--site-install", "--siteinstall",
     action="store_true",
-    help="set up coconut.convenience to be imported on Python start",
+    help="set up coconut.api to be imported on Python start",
 )
 
 arguments.add_argument(
