+++ coconut/command/resources/zcoconut.pth
@@ -1,1 +1,1 @@
-import coconut.convenience
+import coconut.api
