+++ coconut/compiler/compiler.py
@@ -452,7 +452,7 @@ class Compiler(Grammar, pickleable_obj):
         """Creates a new compiler with the given parsing parameters."""
         self.setup(*args, **kwargs)
 
-    # changes here should be reflected in __reduce__ and in the stub for coconut.convenience.setup
+    # changes here should be reflected in __reduce__ and in the stub for coconut.api.setup
     def setup(self, target=None, strict=False, minify=False, line_numbers=False, keep_lines=False, no_tco=False, no_wrap=False):
         """Initializes parsing parameters."""
         if target is None:
