+++ coconut/compiler/header.py
@@ -305,8 +305,8 @@ for items in _coconut.iter(_coconut.zip_longest(*self.iters, fillvalue=_coconut_
     return pattern_prepender'''
             if not strict else
             r'''def prepattern(*args, **kwargs):
-    """Deprecated built-in 'prepattern' disabled by --strict compilation; use 'addpattern' instead."""
-    raise _coconut.NameError("deprecated built-in 'prepattern' disabled by --strict compilation; use 'addpattern' instead")'''
+    """Deprecated Coconut built-in 'prepattern' disabled by --strict compilation; use 'addpattern' instead."""
+    raise _coconut.NameError("deprecated Coconut built-in 'prepattern' disabled by --strict compilation; use 'addpattern' instead")'''
         ),
         def_datamaker=(
             r'''def datamaker(data_type):
@@ -314,14 +314,14 @@ for items in _coconut.iter(_coconut.zip_longest(*self.iters, fillvalue=_coconut_
     return _coconut.functools.partial(makedata, data_type)'''
             if not strict else
             r'''def datamaker(*args, **kwargs):
-    """Deprecated built-in 'datamaker' disabled by --strict compilation; use 'makedata' instead."""
-    raise _coconut.NameError("deprecated built-in 'datamaker' disabled by --strict compilation; use 'makedata' instead")'''
+    """Deprecated Coconut built-in 'datamaker' disabled by --strict compilation; use 'makedata' instead."""
+    raise _coconut.NameError("deprecated Coconut built-in 'datamaker' disabled by --strict compilation; use 'makedata' instead")'''
         ),
         of_is_call=(
             "of = call" if not strict else
             r'''def of(*args, **kwargs):
-    """Deprecated built-in 'of' disabled by --strict compilation; use 'call' instead."""
-    raise _coconut.NameError("deprecated built-in 'of' disabled by --strict compilation; use 'call' instead")'''
+    """Deprecated Coconut built-in 'of' disabled by --strict compilation; use 'call' instead."""
+    raise _coconut.NameError("deprecated Coconut built-in 'of' disabled by --strict compilation; use 'call' instead")'''
         ),
         return_method_of_self=pycondition(
             (3,),
