+++ coconut/tests/src/extras.coco
@@ -281,14 +281,14 @@ def test_convenience() -> bool:
     assert parse("abc", "lenient") == "abc  #1: abc"
 
     setup()
-    assert "Deprecated built-in 'prepattern' disabled by --strict compilation" not in parse("\n", mode="file")
-    assert "Deprecated built-in 'datamaker' disabled by --strict compilation" not in parse("\n", mode="file")
-    assert "Deprecated built-in 'of' disabled by --strict compilation" not in parse("\n", mode="file")
+    assert "Deprecated Coconut built-in 'prepattern' disabled by --strict compilation" not in parse("\n", mode="file")
+    assert "Deprecated Coconut built-in 'datamaker' disabled by --strict compilation" not in parse("\n", mode="file")
+    assert "Deprecated Coconut built-in 'of' disabled by --strict compilation" not in parse("\n", mode="file")
 
     setup(strict=True)
-    assert "Deprecated built-in 'prepattern' disabled by --strict compilation" in parse("\n", mode="file")
-    assert "Deprecated built-in 'datamaker' disabled by --strict compilation" in parse("\n", mode="file")
-    assert "Deprecated built-in 'of' disabled by --strict compilation" in parse("\n", mode="file")
+    assert "Deprecated Coconut built-in 'prepattern' disabled by --strict compilation" in parse("\n", mode="file")
+    assert "Deprecated Coconut built-in 'datamaker' disabled by --strict compilation" in parse("\n", mode="file")
+    assert "Deprecated Coconut built-in 'of' disabled by --strict compilation" in parse("\n", mode="file")
 
     assert_raises(-> parse("def f(x):\n \t pass"), CoconutStyleError)
     assert_raises(-> parse("lambda x: x"), CoconutStyleError)
